Test for transport before sending to it
A device can unregister at any time and clients have to do a lot of
bookkeeping to avoid warnings because they sent to a device that just
disappeared.

This change allows clients to send messages into the void left by unregister.

It is a bit naive as the resend/unregister logic is still invoked for each
unsent message but no harm should be caused by this extra work.