Rework some cheesy code
Using print() is unsuitable for an API.

There is no way to lose a UDP connection and anyway, we should not close the
event loop as others could still be using it.

The cleanup method is using an alight parameter that cannot be passed in.
Instead, let us close the transport and stop discovery so clients have a way
to free the socket.