Reuse the discovery transport for devices
This makes aiolifx allocate just a single UDP port rather than a port for each
bulb. The discovery object caches light objects (not their task objects) and
delivers incoming datagrams to the relevant light.

Devices then no longer have to care about closing their transport. The device
can stay around and it can be re-registered without leaking file descriptors.

Thus we get the ability to communicate with lights that are unregistered.
Because UDP is unreliable and unconnected, this actually makes sense.

For example, a client can use the unregister event to start a relatively fast
polling. This polling will make the light register again as soon as the first
response is received. It would be up to the client to decide when the light is
considered completely dead and stop the polling again.

The important change is that this fast polling can be made without the
broadcasts that a decreased discovery_interval would impose.

As aiolifx no longer creates sockets for each light, this change removes the
IPv6 half-support. Clients can still achieve the effect by doing something
like this in their register method:

    bulb.inet_addr = (mac_to_ipv6_linklocal(bulb.mac_addr), bulb.inet_addr[1])